solo: Replace custom timekeeping with `clock_gettime`.
This should improve both performance and accuracy since the kernel is
in charge of the ticks.

+++ src/Misc/SynthEngine.h
@@ -309,7 +309,7 @@ class SynthEngine
         void( *guiClosedCallback)(void*);
         void *guiCallbackArg;
 
-        int CHtimer;
+        int64_t CHtimer;
 
         int64_t LFOtime; // used by Pcontinous without Pbpm
         float songBeat; // used by Pbpm without Pcontinous
